requirements.txt
langchain
PyPDF2
beautifulsoup4
Pillow
transformers